function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-storage-blob==12.22.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.22
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2

